+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Fix CI caching across version updates.
 - Added unit testing for EnumTypeWrapper.Value usage
+- Reexport of top level enum values when using `import public`
 
 ## 2.2